HF_frontend/requirements_frontend.txt
pillow
altair
pandas
streamlit
transformers
torch
